// src/convex/schema.ts
export const CHANNEL_TYPE = {
  GROUP: "group",
  DIRECT: "direct",
  THREAD: "thread",
} as const;